refactor :move directory structure test to src
