+++ src/test/parsing.test.js
@@ -1,4 +1,4 @@
-import { parseInputs } from '../../src/parsers/Parsing.js';
+import { parseInputs } from '../parsers/Parsing.js';
 
 describe('parsing test', () => {
   test('만욱,노미,재걸,동호', () => {
